requirements.txt
docker-py
pyyaml
pytimeparse
six
